requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
dj-database-url==0.5.0
Django==3.2.3
django-bootstrap-v5==1.0.3
django-environ==0.4.5
django-heroku==0.3.1
django-phonenumber-field==5.1.0
django-tables2==2.3.4
environ==1.0
gunicorn==20.1.0
importlib-metadata==2.1.1
Pillow==8.2.0
psycopg2==2.8.6
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
whitenoise==5.2.0
zipp==3.4.1
